tm/delete: do not delete shared(ro-opened) drbd diskless resources
CDROM devices are shared as read-only opened (secondary) devices
between VM's and we would always try to delete them when the first
VM was removed, but this would fail if another VM was running on the
same host as it still had openers.

Now we check if the resources still has openers,
before trying to deactivate it.